# ui/kore-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kore API provides the frontend API for the Appvia Kore (kore.appvia.io)",
    "title": "Appvia Kore API",
    "contact": {
      "name": "Appvia Ltd",
      "url": "https://appvia.io",
      "email": "info@appvia.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1alpha1/planpolicies": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Returns all the plan policies",
        "operationId": "ListPlanPolicies",
        "parameters": [
          {
            "type": "string",
            "description": "Returns all plan policies for a specific resource type",
            "name": "kind",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all the plan policies in the kore",
            "schema": {
              "$ref": "#/definitions/v1.PlanPolicyList"
            }
          },
          "default": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      }
    },
    "/api/v1alpha1/planpolicies/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Returns a specific plan policy from the kore",
        "operationId": "GetPlanPolicy",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan policy you wish to retrieve",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan policy definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.PlanPolicy"
            }
          },
          "default": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to create or update a plan policy in the kore",
        "operationId": "UpdatePlanPolicy",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan policy you wish to update",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "description": "The specification for the plan policy you are updating",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1.PlanPolicy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan policy definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.PlanPolicy"
            }
          },
          "default": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to delete a plan policy from the kore",
        "operationId": "RemovePlanPolicy",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan policy you wish to delete",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan policy definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.PlanPolicy"
            }
          },
          "default": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      }
    },
    "/api/v1alpha1/plans": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Returns all the classes available to initialized in the kore",
        "operationId": "ListPlans",
        "parameters": [
          {
            "type": "string",
            "description": "Returns all plans for a specific resource type",
            "name": "kind",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all the plans",
            "schema": {
              "$ref": "#/definitions/v1.PlanList"
            }
          },
          "401": {
            "description": "If not authenticated"
          },
          "403": {
            "description": "If authenticated but not authorized"
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      }
    },
    "/api/v1alpha1/plans/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Returns a specific class plan from the kore",
        "operationId": "GetPlan",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan you wish to retrieve",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan definition",
            "schema": {
              "$ref": "#/definitions/v1.Plan"
            }
          },
          "401": {
            "description": "If not authenticated"
          },
          "403": {
            "description": "If authenticated but not authorized"
          },
          "404": {
            "description": "the plan with the given name doesn't exist"
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to create or update a plan in the kore",
        "operationId": "UpdatePlan",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan you wish to create or update",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "description": "The specification for the plan you are creating or updating",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1.Plan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan definition",
            "schema": {
              "$ref": "#/definitions/v1.Plan"
            }
          },
          "400": {
            "description": "Validation error of supplied parameters/body",
            "schema": {
              "$ref": "#/definitions/validation.Error"
            }
          },
          "401": {
            "description": "If not authenticated"
          },
          "403": {
            "description": "If authenticated but not authorized"
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to delete a plan from the kore",
        "operationId": "RemovePlan",
        "parameters": [
          {
            "type": "string",
            "description": "The name of the plan you wish to delete",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the plan definition",
            "schema": {
              "$ref": "#/definitions/v1.Plan"
            }
          },
          "400": {
            "description": "Validation error of supplied parameters/body",
            "schema": {
              "$ref": "#/definitions/validation.Error"
            }
          },
          "401": {
            "description": "If not authenticated"
          },
          "403": {
            "description": "If authenticated but not authorized"
          },
          "404": {
            "description": "the plan with the given name doesn't exist"
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      }
    },
...
    "/api/v1alpha1/teams/{team}/allocations/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to return an allocation within the team",
        "operationId": "GetAllocation",
        "parameters": [
          {
            "type": "string",
            "description": "Is the name of the team you are acting within",
            "name": "team",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Is the name of the allocation you wish to return",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the former team definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.Allocation"
            }
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Used to create/update an allocation within the team.",
        "operationId": "UpdateAllocation",
        "parameters": [
          {
            "type": "string",
            "description": "Is the name of the team you are acting within",
            "name": "team",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Is the name of the allocation you wish to update",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the former team definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.Allocation"
            }
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Remove an allocation from a team",
        "operationId": "RemoveAllocation",
        "parameters": [
          {
            "type": "string",
            "description": "Is the name of the team you are acting within",
            "name": "team",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Is the name of the allocation you wish to delete",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contains the former team definition from the kore",
            "schema": {
              "$ref": "#/definitions/v1.Allocation"
            }
          },
          "500": {
            "description": "A generic API error containing the cause of the error",
            "schema": {
              "$ref": "#/definitions/apiserver.Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "apiserver.Error": {
      "required": [
        "code",
        "detail",
        "message",
        "uri",
        "verb"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "uri": {
          "type": "string"
        },
        "verb": {
          "type": "string"
        }
      }
    },
...
    "v1.Allocation": {
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1.ObjectMeta"
        },
        "spec": {
          "$ref": "#/definitions/v1.AllocationSpec"
        },
        "status": {
          "$ref": "#/definitions/v1.AllocationStatus"
        }
      }
    },
...
    "v1.AllocationSpec": {
      "required": [
        "name",
        "summary",
        "resource",
        "teams"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "resource": {
          "$ref": "#/definitions/v1.Ownership"
        },
        "summary": {
          "type": "string"
        },
        "teams": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1.AllocationStatus": {
      "properties": {
        "conditions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.Condition"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "v1.Condition": {
      "required": [
        "message",
        "detail"
      ],
      "properties": {
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "v1.ListMeta": {
      "description": "ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.",
      "properties": {
        "continue": {
          "description": "continue may be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.",
          "type": "string"
        },
        "remainingItemCount": {
          "description": "remainingItemCount is the number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.",
          "type": "integer",
          "format": "int64"
        },
        "resourceVersion": {
          "description": "String that identifies the server's internal version of this object that can be used by clients to determine when objects have changed. Value must be treated as opaque by clients and passed unmodified back to the server. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency",
          "type": "string"
        },
        "selfLink": {
          "description": "selfLink is a URL representing this object. Populated by the system. Read-only.\n\nDEPRECATED Kubernetes will stop propagating this field in 1.20 release and the field is planned to be removed in 1.21 release.",
          "type": "string"
        }
      }
    },
...
    "v1.Ownership": {
      "required": [
        "group",
        "version",
        "kind",
        "namespace",
        "name"
      ],
      "properties": {
        "group": {
          "type": "string"
        },
        "kind": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
    "v1.Plan": {
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1.ObjectMeta"
        },
        "spec": {
          "$ref": "#/definitions/v1.PlanSpec"
        },
        "status": {
          "$ref": "#/definitions/v1.PlanStatus"
        }
      }
    },
    "v1.PlanList": {
      "required": [
        "items"
      ],
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.Plan"
          }
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1.ListMeta"
        }
      }
    },
    "v1.PlanPolicy": {
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1.ObjectMeta"
        },
        "spec": {
          "$ref": "#/definitions/v1.PlanPolicySpec"
        },
        "status": {
          "$ref": "#/definitions/v1.PlanPolicyStatus"
        }
      }
    },
    "v1.PlanPolicyList": {
      "required": [
        "items"
      ],
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.PlanPolicy"
          }
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "metadata": {
          "$ref": "#/definitions/v1.ListMeta"
        }
      }
    },
    "v1.PlanPolicyProperty": {
      "required": [
        "name",
        "allowUpdate",
        "disallowUpdate"
      ],
      "properties": {
        "allowUpdate": {
          "type": "boolean"
        },
        "disallowUpdate": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "v1.PlanPolicySpec": {
      "required": [
        "kind",
        "summary",
        "description",
        "properties"
      ],
      "properties": {
        "description": {
          "type": "string"
        },
        "kind": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "properties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.PlanPolicyProperty"
          }
        },
        "summary": {
          "type": "string"
        }
      }
    },
    "v1.PlanPolicyStatus": {
      "required": [
        "conditions",
        "status"
      ],
      "properties": {
        "conditions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.Condition"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
    "v1.PlanSpec": {
      "required": [
        "kind",
        "description",
        "summary",
        "configuration"
      ],
      "properties": {
        "configuration": {
          "type": "object"
        },
        "description": {
          "type": "string"
        },
        "kind": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "summary": {
          "type": "string"
        }
      }
    },
    "v1.PlanStatus": {
      "required": [
        "conditions",
        "status"
      ],
      "properties": {
        "conditions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.Condition"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "validation.Error": {
      "required": [
        "code",
        "message",
        "fieldErrors"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "fieldErrors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/validation.FieldError"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "validation.FieldError": {
      "required": [
        "field",
        "errCode",
        "message"
      ],
      "properties": {
        "errCode": {
          "type": "string"
        },
        "field": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "OAuth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "http://localhost:10080/auth",
      "tokenUrl": "http://localhost:10080/token",
      "scopes": {
        "admin": "Admin scope",
        "team": "Team scope"
      }
    }
  },
  "security": [
    {
      "OAuth2": [
        "admin",
        "team"
      ]
    }
  ]
}